Add files via upload
At this update  I have added bootstrap to all of the pages and linked them so the travels is easier but also at the Attendace dropdown, in the Senate page I have corrected table 2, where the 10% from  the missed votes where missing 3 members due to initial code which was  10% from the length of the array. I have also refactored the code for populating table 2 and 3 as long as fixed table 1 where I was printing unnecessary data.