Add files via upload
In the Senate Attendance page  I have refactored the the code which was outputting the values in the table 1, and more specific the column which represented  the percentage for each Party (the last entire column).
I have also added the links with the homepages for each member for the 2nd and the 3rd table at the Attendance Senate page.
The javascript which I have used for building the tables for the Senate Attendance page was also used for generating the tables for the House Attendance page but with data content from the congress house instead, so different values and percentages for the table.
At the House Attendance page  I have encountered NaN values when I was  building the tables, therefore I have inserted if statements before the actual math calculations at line 41, 79 and 88 at the senate-house-attendance.js.
Moving forward I have started to work at the loyalty house and senate pages where there are quite but with different values for the tables. Here I intend to use a different and a more efficient approach.