Add files via upload
I have managed to clear the code when getting the values for the table, fixed the first table which was not generating the accurate values and I have managed to call all the tables in one js file for different html pages using if condition's at the bottom of the "senate-house-attendance-party-loyalty.js" to fix the  null value, so page so the js file knows where to call them. Also I have added if statements in the generating table loop to display in each page the right table headers with the proper values.   Also refactored the function for generating the table. Added a slight styling and will continue to finish the project which is getting close to an end. Looking forward for more !